refactor: rename fetch trains button to animated button
